requirements.txt
flask
flask-cors
pillow
opencv-python-headless
scikit-learn
numpy
